fetch: Added IAuthorizationService and AuthorizationService and adjusted Controllers

+++ TodoList.WebApi/Controllers/TaskAssignmentsController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Mvc.Formatters;
 using TodoList.Services.Enums;
 using TodoList.Services.Interfaces;
 using TodoList.WebApi.Models.Models;
+using IAuthorizationService = TodoList.Services.Interfaces.IAuthorizationService;
 
 namespace TodoList.WebApi.Controllers;
 
@@ -13,10 +14,14 @@ namespace TodoList.WebApi.Controllers;
 public class TaskAssignmentsController : ControllerBase
 {
     private readonly ITaskAssignmentRepository service;
+    private readonly IAuthorizationService authorizationService;
+    private readonly ICurrentUserService currentUserService;
 
-    public TaskAssignmentsController(ITaskAssignmentRepository service)
+    public TaskAssignmentsController(ITaskAssignmentRepository service, Services.Interfaces.IAuthorizationService authorizationService, ICurrentUserService currentUserService)
     {
         this.service = service;
+        this.authorizationService = authorizationService;
+        this.currentUserService = currentUserService;
     }
 
     [HttpGet]
@@ -29,6 +34,20 @@ public class TaskAssignmentsController : ControllerBase
     [HttpDelete("{taskId}/Users/{userId}")]
     public async Task<IActionResult> DeleteAsync(string userId, int taskId)
     {
+        var currentUserId = this.currentUserService.UserId;
+
+        if (userId == null)
+        {
+            throw new UnauthorizedAccessException();
+        }
+
+        var permission = await this.authorizationService.CanEditTasksAsync(currentUserId, taskId);
+
+        if (!permission)
+        {
+            throw new UnauthorizedAccessException();
+        }
+
         await this.service.RemoveTaskAssignmentAsync(userId, taskId);
         return NoContent();
     }
@@ -36,6 +55,20 @@ public class TaskAssignmentsController : ControllerBase
     [HttpPost("{taskId}/Users/{userId}")]
     public async Task<IActionResult> PostAsync(string userId, int taskId)
     {
+        var currentUserId = this.currentUserService.UserId;
+
+        if (userId == null)
+        {
+            throw new UnauthorizedAccessException();
+        }
+
+        var permission = await this.authorizationService.CanEditTasksAsync(currentUserId, taskId);
+
+        if (!permission)
+        {
+            throw new UnauthorizedAccessException();
+        }
+
         await this.service.AssignTaskToUserAsync(userId, taskId);
         return Created();
     }
